Refactored responses and objects. CLI runs command in a new thread.
Refactored responses by renaming them to savers and objects by renaming them to responses, giving the them more logical names.
The commands get run on a new thread so it allows the threading inside the commands to join and finish before running the rest of the command without blocking the CLI and the main thread.